+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 2.0.0 - TBD
 
 ### Changed
-- Use security-context@3.
+- **BREAKING**: Refactor module for use with bedrock@2.
+- **BREAKING**: Use security-context@3.
 
 ## 1.0.0 - 2019-03-29